requirements.txt
pandas>=1.5.0
numpy>=1.21.0
flask>=2.0.0
openpyxl>=3.0.0
pyarrow>=10.0.0
matplotlib>=3.7.0

# browser tests
playwright
pytest-playwright
